refactor: audit and harden 58 files across SDK (R132-R231)
Systematic code quality audit covering unwrap elimination, clippy fixes,
type safety improvements, and dead code removal across all major modules.

Key changes:
- Eliminate unsafe .unwrap()/.expect() in production code paths
- Fix clone_on_copy: add Copy to CacheConfig, remove redundant .clone()
- Type safety: i32→i64 for neo_token votes/gas, u64 as i64→try_from
- Fix PartialEq bug in ProviderError (7 missing match arms)
- Fix unreachable match arms in invocation_result deserializer
- Replace expect_throw with recoverable error in wasm ws/backend
- Simplify invoke_function match block, remove needless return
- Use matches!() macro, .values() iterator, into_owned() idioms
- Remove unused imports in monitoring/metrics and monitoring/tracing
- Add Default impl for SecureStorage, fix non-existent SgxError variant
- Safer float-to-u64 ceiling, lossless u32→u64 via From trait

+++ src/neo_contract/traits/nft.rs
@@ -21,7 +21,7 @@ pub trait NonFungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> + Sen
 
 	// Token methods
 
-	async fn balance_of(&mut self, owner: H160) -> Result<i32, ContractError> {
+	async fn balance_of(&mut self, owner: H160) -> Result<i64, ContractError> {
 		self.call_function_returning_int(
 			<NftContract<P> as NonFungibleTokenTrait<P>>::BALANCE_OF,
 			vec![owner.into()],
@@ -298,7 +298,7 @@ pub trait NonFungibleTokenTrait<'a, P: JsonRpcProvider>: TokenTrait<'a, P> + Sen
 		&mut self,
 		owner: H160,
 		token_id: Bytes,
-	) -> Result<i32, ContractError> {
+	) -> Result<i64, ContractError> {
 		self.throw_if_non_divisible_nft().await?;
 
 		self.call_function_returning_int(
